% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{geodata}
\alias{geodata}
\docType{data}
\title{
Data untuk packages geometric
}
\description{
pada data ini terdapat 3 buah data yaitu data1, data2, dan data3.
}
\usage{data("geodata")}
\format{
  The format is:
 num [1:7, 1:3] 5 6 6 7 8 9 10 4 6 2 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:3] "data1" "data2" "data3"
}
\details{
data1 adalah data tentang A. data2 adalah data tentang B. data3 adalah data tentang C
}
\references{
Kencana,Saka,2018,Kalkulus2,UIIPress,Yogyakarta
}
\examples{
\dontrun{data(geodata)}
}
\keyword{datasets}

