\name{mean_geo}
\alias{mean_geo}
\title{Mean of Geometric Distribution}
\usage{
mean_geo(p)
}
\arguments{
\item{p}{probability of success}
}
\value{
The mean of geometric
}
\description{
Mean pada distribusi geometrik adalah nilai ekspektasi X atau rata-rata dari distribusi geometrik
}
\examples{
mean1 <- mean_geo(0.6)
}
