\name{density_geo}
\alias{density_geo}
\title{PDF of Geometric Distribution}
\usage{
density_geo(p,x)
}
\arguments{
\item{p}{is the probability}
{x}{is sum of experiment/trial}
}
\value{
The pdf of geometric
}
\description{
Probability Density Function (PDF) atau fungsi kepadatan peluang adalah fungsi yang menunjukkan kemungkinan relatif terjadinya hasil yang berbeda dari variabel.
}
\examples{
pdf1 <- density_geo(0.6,3)
}
