\name{var_geo}
\alias{var_geo}
\title{Variance of Geometric Distribution}
\usage{
var_geo(p)
}
\arguments{
\item{p}{is the probability}
}
\value{
The variance of geometric
}
\description{
Variansi adalah rata-rata ukuran seberapa jauh sebuah kumpulan bilangan tersebar, dalam suatu distribusi juga dihitung nilai variansinya termasuk distribusi geometrik yang dihitung menggunakan fungsi ini.
}
\examples{
variance <- var_geo(0.6)
}
